Fix scrolling issues
Addresses an issue where scrolling the list item into view would cause
mouse events to fire even if the mouse wasn't moved. This would cause
undesire behavior navigating with the keybaord, or when active props
were set and the mouse wasn't in use.

Switches to using the standard `scrollIntoView` with an acceptable
default rather than the non-standard `scrollIntoViewIfNeeded`. The
current behavior is a reasonable tradeoff for not requireing a polyfill.